Moved computer vision functions to vision.h
