+++ include/arrayfire.h
@@ -203,3 +203,4 @@
 #include "af/statistics.h"
 #include "af/timing.h"
 #include "af/util.h"
+#include "af/vision.h"
