+++ src/api/c/orb.cpp
@@ -10,7 +10,7 @@
 #include <af/dim4.hpp>
 #include <af/defines.h>
 #include <af/features.h>
-#include <af/image.h>
+#include <af/vision.h>
 #include <handle.hpp>
 #include <err_common.hpp>
 #include <backend.hpp>
